requirements.txt
torch
pymongo
python-dotenv
